refactor(shared-log): extract announce module
Pure code motion of the replication-state announcement retry/repair worker
and the replicator liveness subsystem out of index.ts, per the stage-1
extraction manifest (two modules, since the subsystems share no state):

- src/replication-announcement.ts: ReplicationAnnouncementCoordinator<R>
  with the retry/repair debounce setup, queue/cancel/advance generation
  logic, the repair worker, sendReplicationAnnouncement ordering barrier,
  and the transient-error classifiers plus their constants.
- src/replicator-liveness.ts: ReplicatorLivenessMonitor with the sweep
  timer, target rotation, activity tracking, ACKed ping probe, subscriber
  presence confirmation, and liveness eviction plus their constants.
- src/errors.ts: isNotStartedError moved here (needed by index.ts and both
  new modules without importing index.ts).

SharedLog holds one instance of each (constructed in the ctor, ??= plus
resetForOpen() in open() so borsh clones and reopens keep today's reset
semantics) and keeps thin private delegators/accessors under the exact
prior names so all existing spec files work unchanged. Boundary access
goes through late-bound arrow deps so sinon stubs on the log instance and
service objects keep intercepting.

Manifest deviations:
- Line numbers re-derived by symbol (prior extractions shifted them).
- probeReplicatorLiveness reaches confirmReplicatorSubscriberPresence via
  a dep routed through the SharedLog instance (not module-internally):
  replicator-liveness.spec overrides that method on the log and expects
  in-flight probes to observe the override.
- Module-internal markReplicatorActivity/queue* calls stay internal (the
  spies on those names only assert not-called via the index.ts paths).
- Ratchet BASELINE shrunk by the three announcement generation fields
  that left index.ts; ratchet TARGET intentionally left index.ts-only.
