+++ scripts/ci/check-fence-ratchet.mjs
@@ -54,9 +54,6 @@ const BASELINE = new Set([
 	"_receiveOwnershipMutationAdmissions",
 	"_receiveOwnershipRevision",
 	"_repairLifecycleController",
-	"_replicationAnnouncementRepairGeneration",
-	"_replicationAnnouncementRepairGenerationController",
-	"_replicationAnnouncementRetryGeneration",
 	"_replicationInfoReceiveEpochByPeer",
 	"_replicationLifecycleController",
 	"_subscriptionEpochByPeer",
